requirement.txt
langchain==0.1.13
youtube-transcript-api==0.6.2 
faiss-cpu==1.8.0
sentence-transformers==2.6.0 
streamlit==1.32.2  
openai==1.14.2 
python-dotenv==1.0.1
wikipedia==1.4.0  
tiktoken==0.6.0 
beautifulsoup4==4.12.3 
pypdf2==3.0.1           